Update solve file organization
