+++ tuprolog/solve/_adapters.py
