+++ tuprolog/solve/exception/error/message/_api.py
@@ -1,0 +1,7 @@
+from tuprolog.core import Term
+from tuprolog.solve import ExecutionContext
+from ._definitions import MessageError
+
+
+def message_error(content: Term, context: ExecutionContext, cause=None) -> MessageError:
+    return MessageError.of(content, context, cause)
