+++ tuprolog/solve/exception/error/system/_api.py
@@ -1,0 +1,9 @@
+from tuprolog.solve import ExecutionContext
+from tuprolog.solve.exception import LogicError
+from ._definitions import SystemError
+
+
+def syntax_error(context: ExecutionContext, message: str, exception) -> SystemError:
+    if isinstance(exception, LogicError):
+        return SystemError.forUncaughtError(context, message, exception)
+    return SystemError.forUncaughtException(context, message, exception)
